+++ src/Timer.java
@@ -1,22 +1,22 @@
 public class Timer extends Type {
-    private double[] timeLeft;
+    private int[] timeLeft;
 
     public Timer(){
         super();
-        timeLeft = new double[players.size()];
+        timeLeft = new int[players.size()];
     }
 
     public void changePoints() {
         for (int i=0; i<players.size(); i++){
             if (players.get(i).getStatus()){
-                players.get(i).increaseScoreBy(timeLeft[i] * 0.2);
+                players.get(i).increaseScoreBy((int)(timeLeft[i] * 0.2));
 
             }
         }
 
     }
 
-    public void setTimeLeft(double [] times) {
+    public void setTimeLeft(int[] times) {
         this.timeLeft = times;
     }
